refactor: use to_boolean when comparing boolean vars
When comparing user-defined boolean variables (e.g. ENABLE_<FOO>),
a comparison against the string "0" or "1" is error-prone, as the user
might use other strings which also denote True / False. For that,
bitbake offers the bb.utils.to_boolean function.

This patch refactors all these patterns which are in scope of ISAR.
The parts which are copied from OE are not touched, as well as the WIC
plugins.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>