sdk: Fix build with custom DISTRO when host==target
HOST_DISTRO is not always the right choice for building an SDK. If the
host arch is identical to the target and DISTRO was customized, there is
no HOST_DISTRO to get the bootstrap from. Account for that.

This issue could have been reproduced e.g by xenomai-images when
selecting an x86 target and requesting to build the SDK for it.

Reported-by: Fabian Scheler <fabian.scheler@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>