do not log clamped file mtimes
Modified file timestamps are expected in Debian and need to be clamped
to a fixed value to get reproducible builds. Explicitly logging which
files were touched does not add any value, as these lists are huge and
not related to any packages. By that, just don't log it.

This change is in sync with the reproducible debian installer, as well
as the rootfs handling in OE.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>