patch: Set dummy credentials for git notes
lib/oe/patch was migrated to use git notes to store the filenames for
the patches. This requires git to be configured with some credentials.
To keep compatibility with downstreams and do not introduce any new
requirements for Isar users add some dummy value to be used internally.
This patch will probably need to be reapplied on next Bitbake update.

Signed-off-by: Anton Mikanovich <amikan@ilbers.de>