wic: Restore compatibility with old distros
After update of WIC it requires sfdisk to support --disk-id command,
which is false for Debian Buster and Ubuntu Focal.
Do not try to set disk identifier for those distro versions to keep
them buildable but with lack of reproducibility.

This hack should be removed after deprecation of Buster and Focal.

Signed-off-by: Anton Mikanovich <amikan@ilbers.de>