dpkg-source: Build source package only once
Avoid building the source package multiple times, possibly even
inconsistently. This is achieved by delegating this task to to the base
package and installing the source package from isar-apt in the native
and compat package variants. Those derived packages will also no longer
trigger fetching, unpacking or prepare_build as everything is supposed
to be in the source package.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>