doc/user_manual.md: update ISAR_CROSS_COMPILE variable in SDK part
With the present documentation the SDK is not generated properly for
foreign architectures, basically SDK_TOOLCAHIN not picking the proper
toolchain due to mode is not properly set for foreign architecture.

Set ISAR_CROSS_COMPILE by 1 in local.conf to get the correct mode for
foreign architecture.

Update the path of scripts instead of relative path for accessing files
from isar directory.

Signed-off-by: Srinuvasan A <srinuvasan.a@siemens.com>