sdk-files/files/relocate-sdk.sh: find only ELF executable before apply patchelf
With the present implementation find statement return the regular
executable files under bin,sbin and usr/lib/gcc* which contains the
script files as well.

Actually we are tuning interp and rpath only for ELF executable, hence
added additional condition in the find statement to return only ELF
executable.

With this change we are avoiding unnecessarily checking the interpreter
for non ELF executable files.

Signed-off-by: Srinuvasan A <srinuvasan.a@siemens.com>