meta: Add option to specify additional dependencies for package expand-on-first-boot in case an encrypted disk has to be resized
Although, disk encryption itself is not handled in isar (as of now), downstream projects
using disk encryption break this package. To mitigate this, we added an additional variable
to specify debian packages to be installed, when disk encryption is used. Depending on the
specific implementation (tpm, password prompt, ...) downstream projects can set the dependencies
needed for their strategy to handle the resize.

Note: We decided against fixing this by bbappending in downstream packages, because we believe
this approach scales better and is also easier to maintain.

Signed-off-by: Alexander Heinisch <alexander.heinisch@siemens.com>
Reviewed-by: Jan Kiszka <jan.kiszka@siemes.com>