dpkg: use BPN instead of PN when defining PP
With 2ca3a7e5, sources are built only once and re-used when building
-native and -compat binary packages. The dpkg-raw class generates
a debian/rules file with ${PP}/image expanded: this causes build
of the -compat or -native variants to fail since /home/builder/${PN}
does not exist in the sbuild environment of -compat/-native.

Change the definition of PP to be /home/builder/${BPN} so that the
expanded value remains the same for all variants.

Having absolute paths in generated debian/rules files is obviously
questionable.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>