meta-isar: Added unattended mode for deploy-image-wic installer script.
In some cases (e.g. automated mass rollouts for precomissioning devices)
it is valueable to not rely on a screen and a keyboard to deploy the
target images on the device.

Therefore, extended on the auto.install patch to allow for completely
unattended installations.

In addition to the auto.install file to provide configuration for the installer we added support
for the kernel cmdline.
 - installer.unattended
 - installer.image.uri ...file name of the image to be installed (parameter name uri chosen since we plan to support download of images in upcoming patches)
 - installer.target.dev ...target device name (e.g. /dev/sda) for the image to be installed to
 - installer.target.overwrite ...strategy how to handle target devices not empty (possible values: OVERWRITE - overwrite data on target | ABORT - abort installation if target not empty)

This introduces a **Breaking change!**

This patch changes behaviour since when auto.install is found
the script is assumed to run unattended. That means no dialogs are shown and no
user interaction is enforced! It also requires all the configurations to be in place
and valid when such file is provided!

Signed-off-by: Alexander Heinisch <alexander.heinisch@siemens.com>