add new machine qemuamd64-cip and corresponding mc
The newly added machine qemuamd64-cip provides to build the
kernel from the cip git repository, this would be helpful to demonstrate
the kernel built from git repo rather than tarball.

Moreover, these changes needed to demonstrate the support of CONFIG_LOCALVERSION_AUTO.
in defconfig to automatically append the version in kernelrelease.

We chose this git repo method because the repository should contain
the.git folder to get the latet HEAD revision to append into the kernel release.

Note: The qemuamd64 machine no longer depends on the cip kernel, use qemuamd64-cip
machine to build the images with cip-kernel and updated the KERNEL_DEFCONFIG accordingly.

Signed-off-by: srinuvasan <srinuvasan.a@siemens.com>