linux-custom: Add a provision to support CONFIG_LOCALVERSION_AUTO
This change might be useful for downstream layers to set CONFIG_LOCALVERSION_AUTO
in their base defconfig to append the version automatically to the kernel release.

Presently ISAR doesn't have this support, hence add this provision.

Here we try to find the CONFIG_LOCALVERSION_AUTO availability and .git present during
build the kernel package, unfortunately we are not getting the .git directory in the dpkg_build stage
even if we are fetching the git repo, this is due to sbuild always using the sources (dsc) file
to build the package, when we generate the dsc file it creates the tar ball without .git, hence
during the dpkg_build stage it is not possible to get the latest commit HEAD value using "git rev-parse --verify HEAD"

To solve this issue, we have base .config and .git available in the do_dpkg_source stage, here we try to find the
latest commit and write them into .scmversion file, later during the kernel build package stage
we are calling the kernelrelease target to set the localversion, this setlocalversion script
handle the availability of .scmversion file and append the string accordingly to the kernelrelease.

Added qemuamd64-cip BSP to demonstrate this feature.

Signed-off-by: srinuvasan <srinuvasan.a@siemens.com>