testsuite: Don't call bitbake twice when VM test is run
When running qemu machine with `vm_start()`, CIUtils.getVars() (e.g.
bitbake) is called twice:
 1. from vm_turn_on.format_qemu_cmdline();
 2. from vm_parse_output().

Sometimes, when the second call is immediatelly followed by the first
call of the next test, bitbake daemon gets restarted and it may lead
to an internal avocado error causing cancelling the rest of the tests.

With this patch, we store required bitbake env variables in vm context
and reuse them in the end of the test.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>