Revert "meta: Add option to specify additional dependencies for package expand-on-first-boot in case an encrypted disk has to be resized"
This reverts commit 8b30a4f86cb3ea3369bff3884141872c3a7d9979.

On second thought, this approach turned out to be inapplicable on the
long-run. It is built around the assumption that the disk encryption
secret is still accessible after initramfs used it to unload the disk.
While the downstream implementation of cip-core currently fulfills this,
it is not expected to stay like that because of the increase attack
surface.

We will need a different solution for expanding encrypted partitions,
most likely with the help of the encryption hook in the initramfs.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>