isar-installer: Fixing supervised image deployment in isar-installer when target device is not empty
In supervised mode (!unattended) the check to override non empty
targets was combined with the check for unattended mode. Thus,
resulting in the wrong condition branch when the user  accepted
an override (closed dialog with 'Yes')
In such case the script was assuming similar behaviour like in
unattended mode, thus evaluating the installer_target_overwrite check,
which in that case is neither set via kernel cmdline nor via
auto.install file.

Signed-off-by: Alexander Heinisch <alexander.heinisch@siemens.com>