initramfs-hook: Add infrastructure to ease writing hooks
This recipe include shall simplify writing of initramfs hooks. It
provides the usual headers for the installation hook as well as the boot
scripts, allow to generate that code that installs executables and
modules into the image, but also supports expanding the scripts with own
snippets or defining own ones completely.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>