initramfs-fsck-ext4-hook: Convert to hook.inc and improve
Use the convention and helping features of the new initramfs hook.inc to
simplify this hook. We move it under recipes-initramfs to align with
other recipes there. Another change is making fsck.ext4 mandatory, both
as dependency and while copying it in via the helper - there is no
reason for this recipe to almost silently ignore any unavailability.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>