repository: align repo_contains_package signature with other repo functions
All repository functions take four arguments:
 - path to the repository
 - path to the database used by the repository management tool
 - codename
 - file/package

repo_contains_package was the only function deviating from this scheme and its
current implementation only takes a repository path and a file. Since the
repositoty class was introduced to hide implementation details (so we could
move away from reprepro if we wanted to), align that function with others.
The only caller found in the repository (base-apt) was changed to use the new
(standard) API.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>