image.bbclass: use oe.utils.directory_size() instead of du
Currently the "du" command is used to calculate the rootfs size, and
sometimes the reported size of the rootfs by "du" varies across multiple
builds (usually by a few kb). But the rootfs contents are reproducible.

Although this does not affect any functionality, but it does affect
reproducibility of artifacts that rely on the calculated rootfs size.

Some of the files/directories in the rootfs has different number of
blocks allocated and "du" estimates the directory size based on the
allocated blocks. The files/directories affected are not always the
same across builds.

OE-Core uses a custom function to calculate directory size (See [1])
and this patch introduces the use of that function from oe.utils to
calculate the directory size.

[1] https://github.com/openembedded/openembedded-core/commit/d8f1f3a6b024a2ae6631d1ce25421e8d94b69a12

Signed-off-by: Adithya Balakumar <Adithya.Balakumar@toshiba-tsip.com>