snapshots: add option to use separate timestamp for security component
Before releasing a product all available security fixes should be
included. However, you might not want to get other proposed updates.
With the previous snapshot logic it was not possible to model this, as a
single timestamp is used for all apt source-list entries.

We change that by adding a "security" flag to snapshot date variables.
By that, dedicated control over the security distribution is possible.

For now, we only add this logic for debian distributions (not ubuntu),
as only there we have a dedicated security distribution.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>