bootstrap: space-separate base packages
The variable DISTRO_BOOTSTRAP_BASE_PACKAGES contains a list of base
packages that need to be included when bootstrapping. This list was
comma separated for technical reasons, which made it hard to append or
remove from that list. Instead, we now make it space separated and join
it by comma where needed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>