make usage of usrmerge distro dependent
The situation around usrmerge is complex: In older distro versions that
must be included to get the /{bin,sbin,lib} dirs symlinked to their
counterparts in /usr. In debian, the essential package
init-system-helpers depends on usrmerge from bookworm on, so it is
discouraged to manually include it (in the future it might be dropped
completely). In Ubuntu, the usrmerge package is already dropped from
noble on, currently breaking the build.

We now add the package using the distro conf just for the distros that
need it, restoring the ubuntu build and making it future-proof.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>