deb-dl-dir: fix package source download
Some of the packages that are part of the bootstrapped base filesystem
were not downloaded when activating "cache-deb-src" via
BASE_REPO_FEATURES.

It seems that some (or all) packages are not listed in
"/var/log/dpkg.log" leading to skip the download of these files.

This is a regression that has been introduced with mmdebstrap, which
either not creates or just deletes the mentioned log file.

Package built by using the dpkg-prebuilt.bbclass tend to reference
binary packages without providing any source package reference.
To handle these, currently valid, cases the download function simply
skips packages that could not be downloaded and prints a warning.

Signed-off-by: Benedikt Niedermayr <benedikt.niedermayr@siemens.com>