linux-distro: add RPROVIDES for distro kernel packages
When building a non-distro kernel, all generated packages are also
registered as RPROVIDES, so that downstream user can runtime depend on
them. This however was not implemented for the distro kernel dummy
packages.

We now add RPROVIDES for all distro kernel packages as well. As this
would lead to a lot of code duplication, we generalize the anonymous
python function a bit.

Co-developed-by: Christoph Steiger <christoph.steiger@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>