doc/user_manual: Fix content
The usermod command checks whether the user is currently running a
process before attempting to change the user's home directory.
This check will always fail for the root user (Error: usermod: user root
is currently used by process <x>). It is therefore not possible to
change the root user's home directory in this way.

Signed-off-by: Clara Kowalsky <clara.kowalsky@siemens.com>