fix: add currenttask-guards to SSTATE(CREATE|POSTINST)FUNCS
We use SSTATECREATEFUNCS and SSTATEPOSTINSTFUNCS to insert
code for preparation/postprocessing of sstate artifacts.
As it turns out, those are not per-job, but global, so if
a recipe has multiple sstate-cacheable tasks, all of those
functions are executed, and we need to make sure to bail out
of all unintended calls.
Because all invocations are visible in the logs, we add a
message in case a function was skipped, to make logs more
readable.

Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>