wic: efi-plus-pcbios: Use --use-uuid for partitions (vs --ondisk sda)
After installing an image from isar-installer-image for x86-pc to an NVMe
drive, the boot process incorrectly checks /dev/sda1 for the boot partition,
as specified in /etc/fstab.
This is not an issue with devices using SATA drives but causes problems when
NVMe drives or multiple SATA drives are present.
This change checks for partition UUIDs vs /dev/sda device.

Signed-off-by: badrikesh prusty <badrikesh.prusty@siemens.com>